Add template for project
